# Security initialization on re-addition