--- a/Algorithm.CSharp/RawPricesUniverseRegressionAlgorithm.cs
+++ b/Algorithm.CSharp/RawPricesUniverseRegressionAlgorithm.cs
@@ -91,12 +91,12 @@ namespace QuantConnect.Algorithm.CSharp
         /// <summary>
         /// Data Points count of all timeslices of algorithm
         /// </summary>
-        public long DataPoints => 135;
+        public long DataPoints => 156;
 
         /// <summary>
         /// Data Points count of the algorithm history
         /// </summary>
-        public int AlgorithmHistoryDataPoints => 30;
+        public int AlgorithmHistoryDataPoints => 150;
 
         /// <summary>
         /// Final status of the algorithm
@@ -108,33 +108,33 @@ namespace QuantConnect.Algorithm.CSharp
         /// </summary>
         public Dictionary<string, string> ExpectedStatistics => new Dictionary<string, string>
         {
-            {"Total Orders", "12"},
-            {"Average Win", "0.34%"},
-            {"Average Loss", "-0.14%"},
-            {"Compounding Annual Return", "4.586%"},
-            {"Drawdown", "0.700%"},
-            {"Expectancy", "0.158"},
+            {"Total Orders", "57"},
+            {"Average Win", "0.18%"},
+            {"Average Loss", "-0.24%"},
+            {"Compounding Annual Return", "-47.380%"},
+            {"Drawdown", "2.500%"},
+            {"Expectancy", "-0.352"},
             {"Start Equity", "50000"},
-            {"End Equity", "50090.17"},
-            {"Net Profit", "0.180%"},
-            {"Sharpe Ratio", "5.991"},
-            {"Sortino Ratio", "0"},
-            {"Probabilistic Sharpe Ratio", "99.393%"},
-            {"Loss Rate", "67%"},
-            {"Win Rate", "33%"},
-            {"Profit-Loss Ratio", "2.47"},
-            {"Alpha", "0.17"},
-            {"Beta", "0.029"},
-            {"Annual Standard Deviation", "0.028"},
-            {"Annual Variance", "0.001"},
-            {"Information Ratio", "2.734"},
-            {"Tracking Error", "0.098"},
-            {"Treynor Ratio", "5.803"},
+            {"End Equity", "48726.48"},
+            {"Net Profit", "-2.547%"},
+            {"Sharpe Ratio", "-3.372"},
+            {"Sortino Ratio", "-3.889"},
+            {"Probabilistic Sharpe Ratio", "10.352%"},
+            {"Loss Rate", "63%"},
+            {"Win Rate", "37%"},
+            {"Profit-Loss Ratio", "0.75"},
+            {"Alpha", "-0.208"},
+            {"Beta", "0.815"},
+            {"Annual Standard Deviation", "0.086"},
+            {"Annual Variance", "0.007"},
+            {"Information Ratio", "-4.871"},
+            {"Tracking Error", "0.039"},
+            {"Treynor Ratio", "-0.357"},
             {"Total Fees", "$0.00"},
-            {"Estimated Strategy Capacity", "$99000000.00"},
+            {"Estimated Strategy Capacity", "$230000000.00"},
             {"Lowest Capacity Asset", "AIG R735QTJ8XC9X"},
-            {"Portfolio Turnover", "15.96%"},
-            {"OrderListHash", "d915ae36ce856457b32ebbfce4581281"}
+            {"Portfolio Turnover", "77.40%"},
+            {"OrderListHash", "4fb8ffbdfd2cce69ac28b0d0992d7198"}
         };
     }
 }