<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->

Ensure securities are initialized every time they are added or re-added, including scenarios like manual removal and re-addition and deselection and re-selection by universes, like options or custom universes.
Also make sure initialization happens only once for each security on addition.

##### Performance tests:

Since now the securities are being initialized every time they are re-added (either manually or by being selected by a universe), we expect to have minor performance impacts on algorithms that are doing selection.

**Daily coarse selection: 50 top equities by dollar volume and their options:**

```csharp
public override void Initialize()
{
    SetStartDate(2014, 01, 01);
    SetEndDate(2014, 03, 01);
    SetCash(100000);

    UniverseSettings.Resolution = Resolution.Daily;

    var seeder = new FuncSecuritySeeder(security => GetLastKnownPrices(security));
    SetSecurityInitializer(security => seeder.SeedSecurity(security));

    var universe = AddUniverse(coarse => coarse.OrderByDescending(x => x.DollarVolume).Take(50).Select(x => x.Symbol));

    AddUniverseOptions(universe, u => u.Expiration(0, 7));
}
```

Master: ~136s on avg
PR: ~146s on avg (~7% slower)

**Daily coarse selection: randomly selecting 50 out 200 top equities by dollar volume and their options:**

```csharp
var universe = AddUniverse(coarse => coarse.OrderByDescending(x => x.DollarVolume).Take(200).OrderBy(_ => random.Next()).Take(50).Select(x => x.Symbol));
AddUniverseOptions(universe, u => u.Expiration(0, 7));
```

Note: selecting 50 randomly so we force some equities are de-selected and re-selected each day, which forces more security initializations.

Master: ~166s on avg
PR: ~191s on avg (~7% slower)

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->

Part of #7929 

#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->

#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->

#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

Regression algorithms

#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] Refactor (non-breaking change which improves implementation)
- [ ] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->